test: seo checks
