+++ tests/general.spec.ts
@@ -7,17 +7,6 @@ test.beforeEach(async ({ page }) => {
   })
 })
 
-test("SEO cheks", async ({ page }) => {
-  await test.step("Check the title is correct", async () => {
-    await expect(page).toHaveTitle("Automation Chronicles")
-  })
-
-  await test.step("Check the description is correct", async () => {
-    const description = page.locator("meta[name=description]")
-    await expect(description).toHaveAttribute("content", "Blog zameraný na automatizáciu a vývoj softvéru")
-  })
-})
-
 test("HTML has correct lang attribute", async ({ page }) => {
   await test.step("Check the lang attribute is correct", async () => {
     const html = page.locator("html")
